fix: avoid UMP API on web
