+++ lib/ads/consent_manager_web.dart
@@ -1,0 +1,8 @@
+import 'consent_manager.dart';
+
+Future<void> maybeShowConsentForm({
+  ConsentClient? consentClient,
+  ConsentFormPresenter? presenter,
+}) async {
+  // Consent handling is not required on web.
+}
